# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Lightspeed Core Service (LCS) service - OpenAPI",
        "summary": "Lightspeed Core Service (LCS) service API specification.",
        "description": "Lightspeed Core Service (LCS) service API specification.",
        "contact": {
            "name": "Pavel Tisnovsky",
            "url": "https://github.com/tisnik/",
            "email": "ptisnovs@redhat.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.2.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/",
            "description": "Locally running service"
        }
    ],
    "paths": {
...
        "/v1/feedback": {
            "post": {
                "tags": [
                    "feedback"
                ],
                "summary": "Feedback Endpoint Handler",
                "description": "Handle feedback requests.\n\nProcesses a user feedback submission, storing the feedback and\nreturning a confirmation response.\n\nArgs:\n    feedback_request: The request containing feedback information.\n    ensure_feedback_enabled: The feedback handler (FastAPI Depends) that\n        will handle feedback status checks.\n    auth: The Authentication handler (FastAPI Depends) that will\n        handle authentication Logic.\n\nReturns:\n    Response indicating the status of the feedback storage request.\n\nRaises:\n    HTTPException: Returns HTTP 500 if feedback storage fails.",
                "operationId": "feedback_endpoint_handler_v1_feedback_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Feedback received and stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid credentials provided by client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Client does not have permission to access resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "User feedback can not be stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/feedback/status": {
            "get": {
                "tags": [
                    "feedback"
                ],
                "summary": "Feedback Status",
                "description": "Handle feedback status requests.\n\nReturn the current enabled status of the feedback\nfunctionality.\n\nReturns:\n    StatusResponse: Indicates whether feedback collection is enabled.",
                "operationId": "feedback_status_v1_feedback_status_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "feedback"
                ],
                "summary": "Update Feedback Status",
                "description": "Handle feedback status update requests.\n\nTakes a request with the desired state of the feedback status.\nReturns the updated state of the feedback status based on the request's value.\nThese changes are for the life of the service and are on a per-worker basis.\n\nReturns:\n    FeedbackStatusUpdateResponse: Indicates whether feedback is enabled.",
                "operationId": "update_feedback_status_v1_feedback_status_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackStatusUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackStatusUpdateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "detail": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Detail",
                        "description": "Error details",
                        "examples": [
                            {
                                "cause": "Failed to handle request to https://bam-api.res.ibm.com/v2/text",
                                "response": "Error while validation question"
                            },
                            {
                                "cause": "Invalid conversation ID 1237-e89b-12d3-a456-426614174000",
                                "response": "Error retrieving conversation history"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ErrorResponse",
                "description": "Model representing error response for query endpoint.",
                "examples": [
                    {
                        "detail": {
                            "cause": "Failed to handle request to https://bam-api.res.ibm.com/v2/text",
                            "response": "Error while validation question"
                        }
                    },
                    {
                        "detail": {
                            "cause": "Invalid conversation ID 1237-e89b-12d3-a456-426614174000",
                            "response": "Error retrieving conversation history"
                        }
                    }
                ]
            },
            "FeedbackCategory": {
                "type": "string",
                "enum": [
                    "incorrect",
                    "not_relevant",
                    "incomplete",
                    "outdated_information",
                    "unsafe",
                    "other"
                ],
                "title": "FeedbackCategory",
                "description": "Enum representing predefined feedback categories for AI responses.\n\nThese categories help provide structured feedback about AI inference quality\nwhen users provide negative feedback (thumbs down). Multiple categories can\nbe selected to provide comprehensive feedback about response issues."
            },
            "FeedbackRequest": {
                "properties": {
                    "conversation_id": {
                        "type": "string",
                        "title": "Conversation Id",
                        "description": "The required conversation ID (UUID)",
                        "examples": [
                            "c5260aec-4d82-4370-9fdf-05cf908b3f16"
                        ]
                    },
                    "user_question": {
                        "type": "string",
                        "title": "User Question",
                        "description": "User question (the query string)",
                        "examples": [
                            "What is Kubernetes?"
                        ]
                    },
                    "llm_response": {
                        "type": "string",
                        "title": "Llm Response",
                        "description": "Response from LLM",
                        "examples": [
                            "Kubernetes is an open-source container orchestration system for automating ..."
                        ]
                    },
                    "sentiment": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sentiment",
                        "description": "User sentiment, if provided must be -1 or 1",
                        "examples": [
                            -1,
                            1
                        ]
                    },
                    "user_feedback": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 4096
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Feedback",
                        "description": "Feedback on the LLM response.",
                        "examples": [
                            "I'm not satisfied with the response because it is too vague."
                        ]
                    },
                    "categories": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/FeedbackCategory"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Categories",
                        "description": "List of feedback categories that describe issues with the LLM response (for negative feedback).",
                        "examples": [
                            [
                                "incorrect",
                                "incomplete"
                            ]
                        ]
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "conversation_id",
                    "user_question",
                    "llm_response"
                ],
                "title": "FeedbackRequest",
                "description": "Model representing a feedback request.\n\nAttributes:\n    conversation_id: The required conversation ID (UUID).\n    user_question: The required user question.\n    llm_response: The required LLM response.\n    sentiment: The optional sentiment.\n    user_feedback: The optional user feedback.\n    categories: The optional list of feedback categories (multi-select for negative feedback).\n\nExample:\n    ```python\n    feedback_request = FeedbackRequest(\n        conversation_id=\"12345678-abcd-0000-0123-456789abcdef\",\n        user_question=\"what are you doing?\",\n        user_feedback=\"This response is not helpful\",\n        llm_response=\"I don't know\",\n        sentiment=-1,\n        categories=[FeedbackCategory.INCORRECT, FeedbackCategory.INCOMPLETE]\n    )\n    ```",
                "examples": [
                    {
                        "conversation_id": "12345678-abcd-0000-0123-456789abcdef",
                        "llm_response": "bar",
                        "sentiment": -1,
                        "user_feedback": "Not satisfied with the response quality.",
                        "user_question": "foo"
                    },
                    {
                        "categories": [
                            "incorrect"
                        ],
                        "conversation_id": "12345678-abcd-0000-0123-456789abcdef",
                        "llm_response": "The capital of France is Berlin.",
                        "sentiment": -1,
                        "user_question": "What is the capital of France?"
                    },
                    {
                        "categories": [
                            "incomplete",
                            "not_relevant"
                        ],
                        "conversation_id": "12345678-abcd-0000-0123-456789abcdef",
                        "llm_response": "Use Docker.",
                        "sentiment": -1,
                        "user_feedback": "This response is too general and doesn't provide specific steps.",
                        "user_question": "How do I deploy a web app?"
                    }
                ]
            },
            "FeedbackResponse": {
                "properties": {
                    "response": {
                        "type": "string",
                        "title": "Response"
                    }
                },
                "type": "object",
                "required": [
                    "response"
                ],
                "title": "FeedbackResponse",
                "description": "Model representing a response to a feedback request.\n\nAttributes:\n    response: The response of the feedback request.\n\nExample:\n    ```python\n    feedback_response = FeedbackResponse(response=\"feedback received\")\n    ```",
                "examples": [
                    {
                        "response": "feedback received"
                    }
                ]
            },
            "FeedbackStatusUpdateRequest": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "title": "Status",
                        "description": "Desired state of feedback enablement, must be False or True",
                        "default": false,
                        "examples": [
                            true,
                            false
                        ]
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "title": "FeedbackStatusUpdateRequest",
                "description": "Model representing a feedback status update request.\n\nAttributes:\n    status: Value of the desired feedback enabled state.\n\nExample:\n    ```python\n    feedback_request = FeedbackRequest(\n        status=false\n    )\n    ```"
            },
            "FeedbackStatusUpdateResponse": {
                "properties": {
                    "status": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "FeedbackStatusUpdateResponse",
                "description": "Model representing a response to a feedback status update request.\n\nAttributes:\n    status: The previous and current status of the service and who updated it.\n\nExample:\n    ```python\n    status_response = StatusResponse(\n        status={\n            \"previous_status\": true,\n            \"updated_status\": false,\n            \"updated_by\": \"user/test\",\n            \"timestamp\": \"2023-03-15 12:34:56\"\n        },\n    )\n    ```",
                "examples": [
                    {
                        "status": {
                            "previous_status": true,
                            "timestamp": "2023-03-15 12:34:56",
                            "updated_by": "user/test",
                            "updated_status": false
                        }
                    }
                ]
            },
            "ForbiddenResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ForbiddenResponse",
                "description": "Model representing response for forbidden access.",
                "examples": [
                    {
                        "detail": "Forbidden: User is not authorized to access this resource"
                    }
                ]
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "StatusResponse": {
                "properties": {
                    "functionality": {
                        "type": "string",
                        "title": "Functionality"
                    },
                    "status": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "functionality",
                    "status"
                ],
                "title": "StatusResponse",
                "description": "Model representing a response to a status request.\n\nAttributes:\n    functionality: The functionality of the service.\n    status: The status of the service.\n\nExample:\n    ```python\n    status_response = StatusResponse(\n        functionality=\"feedback\",\n        status={\"enabled\": True},\n    )\n    ```",
                "examples": [
                    {
                        "functionality": "feedback",
                        "status": {
                            "enabled": true
                        }
                    }
                ]
            },
...
            "UnauthorizedResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "UnauthorizedResponse",
                "description": "Model representing response for missing or invalid credentials.",
                "examples": [
                    {
                        "detail": "Unauthorized: No auth header found"
                    }
                ]
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}